requirements.txt
requests==2.23.0
wheel==0.34.2
werkzeug==0.15.5
flask_bcrypt==0.7.1
typing-extensions==3.7.4.2
more_itertools==8.3.0
apache-airflow==1.10.10